refactor(#458): split damkaStore.ts (265 lines) into damkaTypes + damkaLogic
